refactor(frontend): extract SetupPanel to components/SetupPanel.jsx (5/N)
Moves the first-run server-config modal out of App.jsx. ~130 LOC of
self-contained UI: URL + token form, live /ping connection test, save +
reload, disconnect, dual mode (blocking on first run / re-opened from
header).

Also exports _readApiUrl from lib/api.js so SetupPanel can re-read the
current value without re-implementing the localStorage probe.

Build clean. App.jsx down to 3577 LOC.
